fix: break circular dependency between config/projects.tsx and config/project-Item.tsx
Extract shared ProjectItems interface into types/project-items.ts to
resolve the circular import dependency. Update both config files to
import ProjectItems from the new types file.
