+++ types/project-items.ts
@@ -1,0 +1,10 @@
+export interface ProjectItems {
+  projectName: string;
+  description: string;
+  projectImage: string;
+  githubLink: string;
+  liveLink?: string;
+  ytLink?: string;
+  techStack: string[];
+  difficulty: string;
+}
